code comments and small updates
