+++ flash-loan/src/lib.rs
@@ -10,5 +10,7 @@ mod vault {
 }
 
 mod contract;
+mod execution;
+mod storage;
+mod token_utility;
 mod types;
-mod utils;
